Remove long-deprecated eclipse xml (long ago replaced by EclipseWtp): - it had a dedicated block in plugin-maven, which hasn't been in the documentation for a very long time - moved the code to plugin-gradle for now, since plugin-gradle will have a breaking change soon, but not yet
