--- lib/src/main/java/com/diffplug/spotless/xml/package-info.java
@@ -1,7 +1,0 @@
-@ParametersAreNonnullByDefault
-@ReturnValuesAreNonnullByDefault
-package com.diffplug.spotless.xml;
-
-import javax.annotation.ParametersAreNonnullByDefault;
-
-import com.diffplug.spotless.annotations.ReturnValuesAreNonnullByDefault;
