+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/libdeprecated/XmlDefaultsTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.xml;
+package com.diffplug.gradle.spotless.libdeprecated;
 
 import static org.assertj.core.api.Assertions.assertThat;
